pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
Flask = "^2.1.2"
dnspython = "^2.2.1"
pymongo = "^4.1.1"
psutil = "^5.9.1"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus.git"}
requests = "^2.28.1"
Pillow = "^9.2.0"
discord = "^2.0.0"
opencv-python = "^4.6.0"
scikit-image = "^0.19.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"